+++ utilities/js/BREAKPOINTS.js
@@ -5,11 +5,6 @@
 
     const $html = document.documentElement;
 
-    // `--rootWidth` is initially set in `THEME/header.php`
-    new ResizeObserver( entries => {
-        $html.style.setProperty( '--rootWidth', `${$html.clientWidth}` );
-    } ).observe( $html );
-
     window.BREAKPOINTS = {
         up: size => $html.clientWidth >= (BREAKPOINTS[size] || size),
         down: size => $html.clientWidth < (BREAKPOINTS[size] || size),