+++ utilities/utilities.php
@@ -11,6 +11,16 @@ add_action( 'after_setup_theme', function() {
 // enable shortcodes for titles
 add_filter( 'the_title', 'do_shortcode' );
 
+add_action( 'wp_head', function() {
+	// set `--rootWidth` (for CSS' fluid calculation)
+	// run it _first_ in `<head>` to immediately set value ?>
+    <script>(function() {
+            function setRootWidth() { document.documentElement.style
+                .setProperty( '--rootWidth', `${document.documentElement.clientWidth}` ); }
+            setRootWidth() || new ResizeObserver( setRootWidth ).observe( document.documentElement )
+        })();</script>
+<?php }, -1 );
+
 add_action( 'wp_enqueue_scripts', function() {
 	wp_register_script( 'behaviours', UTILITIES_DIRECTORY_URI . '/js/behaviours.js', ['alter'], '2.0.0', true );
 	wp_register_script( 'alter', UTILITIES_DIRECTORY_URI . '/js/alter.js', [], '1.0.0', true );